fix: fix club detail button flicker
