+++ packages/web/src/features/clubDetails/frames/ClubDetailMainFrame.tsx
@@ -2,8 +2,10 @@
 
 import React from "react";
 
+import { RegistrationApplicationStudentStatusEnum } from "@sparcs-clubs/interface/common/enum/registration.enum";
 import styled from "styled-components";
 
+import AsyncBoundary from "@sparcs-clubs/web/common/components/AsyncBoundary";
 import FlexWrapper from "@sparcs-clubs/web/common/components/FlexWrapper";
 import PageHead from "@sparcs-clubs/web/common/components/PageHead";
 import SectionTitle from "@sparcs-clubs/web/common/components/SectionTitle";
@@ -13,6 +15,7 @@ import ClubInfoCard from "@sparcs-clubs/web/features/clubDetails/components/Club
 import PersonInfoCard from "@sparcs-clubs/web/features/clubDetails/components/PersonInfoCard";
 
 import { RegisterInfo } from "../components/RegisterInfo";
+import { useGetMyMemberRegistration } from "../services/getMyMemberRegistration";
 
 import type { ApiClb002ResponseOK } from "@sparcs-clubs/interface/api/club/endpoint/apiClb002";
 
@@ -57,45 +60,77 @@ const ClubDetailMainFrame: React.FC<ClubDetailMainFrameProps> = ({
 }) => {
   const { isLoggedIn } = useAuth();
 
+  const {
+    data: myRegistrationList,
+    isLoading,
+    isError,
+    refetch,
+  } = useGetMyMemberRegistration();
+
+  let isInClub = RegistrationApplicationStudentStatusEnum.Rejected;
+  let isRegistered = false;
+  if (myRegistrationList && myRegistrationList.applies.length > 0) {
+    const thisRegistration = myRegistrationList.applies.find(
+      apply => apply.clubId === club.id,
+    );
+    if (thisRegistration) {
+      isInClub = thisRegistration.applyStatusEnumId;
+      isRegistered = true;
+    } else {
+      isRegistered = false;
+    }
+  }
+
   return (
-    <FlexWrapper direction="column" gap={60}>
-      <PageHead
-        items={[
-          { name: "동아리 목록", path: "/clubs" },
-          { name: club.name_kr, path: `/clubs/${club.id}` },
-        ]}
-        title={club.name_kr}
-        action={
-          isLoggedIn && isRegistrationPeriod && <RegisterInfo club={club} />
-        }
-      />
+    <AsyncBoundary isLoading={isLoading} isError={isError}>
+      <FlexWrapper direction="column" gap={60}>
+        <PageHead
+          items={[
+            { name: "동아리 목록", path: "/clubs" },
+            { name: club.name_kr, path: `/clubs/${club.id}` },
+          ]}
+          title={club.name_kr}
+          action={
+            isLoggedIn &&
+            isRegistrationPeriod && (
+              <RegisterInfo
+                club={club}
+                isInClub={isInClub}
+                isRegistered={isRegistered}
+                refetch={refetch}
+                myRegistrationList={myRegistrationList ?? { applies: [] }}
+              />
+            )
+          }
+        />
 
-      <FlexWrapper direction="column" gap={20}>
-        <SectionTitle size="lg">동아리 정보</SectionTitle>
-        <CardWrapper>
-          <ClubInfoCard club={club} />
-        </CardWrapper>
-      </FlexWrapper>
-
-      <MoreInfoWrapper>
-        <PersonInfoWrapper>
-          <SectionTitle size="lg">인적 사항 </SectionTitle>
+        <FlexWrapper direction="column" gap={20}>
+          <SectionTitle size="lg">동아리 정보</SectionTitle>
           <CardWrapper>
-            <PersonInfoCard
-              club={club}
-              isRegistrationPeriod={isRegistrationPeriod}
-            />
-          </CardWrapper>
-        </PersonInfoWrapper>
-
-        <FlexWrapper direction="column" gap={20} style={{ flex: "1 0 0" }}>
-          <SectionTitle size="lg">동아리 설명</SectionTitle>
-          <CardWrapper>
-            <ClubDetailCard club={club} />
+            <ClubInfoCard club={club} />
           </CardWrapper>
         </FlexWrapper>
-      </MoreInfoWrapper>
-    </FlexWrapper>
+
+        <MoreInfoWrapper>
+          <PersonInfoWrapper>
+            <SectionTitle size="lg">인적 사항 </SectionTitle>
+            <CardWrapper>
+              <PersonInfoCard
+                club={club}
+                isRegistrationPeriod={isRegistrationPeriod}
+              />
+            </CardWrapper>
+          </PersonInfoWrapper>
+
+          <FlexWrapper direction="column" gap={20} style={{ flex: "1 0 0" }}>
+            <SectionTitle size="lg">동아리 설명</SectionTitle>
+            <CardWrapper>
+              <ClubDetailCard club={club} />
+            </CardWrapper>
+          </FlexWrapper>
+        </MoreInfoWrapper>
+      </FlexWrapper>
+    </AsyncBoundary>
   );
 };
